Add support for hfile reader
